Refactor GradScaler for simplified AMP support and remove overflow checks
- Simplified GradScaler implementation by removing overflow checks and related statistics.
- Updated initialization parameters and validation logic.
- Enhanced tensor scaling and unscaling methods for better performance.
- Removed unnecessary methods and attributes related to overflow detection.
- Updated state_dict and load_state_dict methods to reflect changes in attributes.

Improve function handling in base.py for autocasting

- Added check for autocast enabled state before applying tensor casting.
- Prevented recursion in Cast operations during tensor processing.

Adjust conv2d function to improve tensor layout

- Transposed slides tensor to optimize data layout for convolution operations.

Remove gradient overflow checks from optimizers

- Removed gradient overflow detection from Adam, RMSprop, and SGD optimizers to streamline updates.

Optimize tensor data handling in Tensor class

- Modified tensor data type casting to avoid unnecessary copies.

Update package version in setup.py

- Bumped version from 0.5.3 to 0.5.5.

Remove outdated state dict tests

- Deleted test_module_state_dict.py as it contained outdated test cases.

+++ neurograd/tensor.py
@@ -13,7 +13,7 @@ class Tensor:
         if not isinstance(data, xp.ndarray):
             self.data = xp.array(data, dtype=dtype)
         elif dtype is not None:
-            self.data = data.astype(dtype)
+            self.data = data.astype(dtype, copy=False)
         else:
             self.data = data
         self.requires_grad = requires_grad # whether to compute gradients for this tensor
